Feature added: Right Click on entries to create group from selection
Added functionality to allow users to right click to make a new group from the selection. Updated all group creation and group editing to use the CreateGroupAction class, extending SimpleCommand so it would work with the right-click menu.

Getting this functionality extracted into an Action that could be called with the available states stored in stateManager required some pretty large changes and bug fixes related to synchronizing the various lists of active and selected entries.

Additionally, fixed a bug where when new entries were dropped into a list, the displayed number of entries did not update.