Correct option text, position, & hiding behavior.
The new option's presence now based on selected style, with presence
being ensured only if the style is a JStyle. The option now appears
always first from the top, using index based insertion to achieve this.

The text has been updated to reference bibliographic entries as opposed
to citations, which was factually incorrect.