Add setting: always add "Cited on pages" text to JStyles. (#11732)
* Add setting to always add "Cited on pages" text in JStyle Citations.

* Add setting: always add "Cited on pages" text to JStyles.

* Add changelog.

* Add change to Changelog.

* Fix translation key. Ensure checkstyle conformity.

* Correct option text, position, & hiding behavior.

The new option's presence now based on selected style, with presence
being ensured only if the style is a JStyle. The option now appears
always first from the top, using index based insertion to achieve this.

The text has been updated to reference bibliographic entries as opposed
to citations, which was factually incorrect.

* Cited on pages... text now appears.

Reference existing preference in OOBibBase to properly propagate
setting, so that "Cited on pages" text literal will appear when user has
it configured.

Remove accomplished TODOs.

* Shift changelog entry, fix setting string, misc. formatting

* Attempt to update submodules to origin/master state.

* Attempt to fix submodule hashes.

* Re-ignore submodules.

* Fix spacing.

---------

Co-authored-by: Subhramit Basu Bhowmick <subhramit.bb@live.in>