Preventing library disappearing when closing tab (#12909)
* Preventing library disappearing when closing tab

If you have two databases and the welcome tab open, then close the tab directly to the right of the welcome tab.

The top bar will be hidden resulting in you "losing" the other database.

* fixing style violation

* Update CHANGELOG.md

* Update JabRefFrame.java

* Fixed bug with home page & closing library tabs

Now if a user closes a tab beside the Home Page tab and there are other libraries active it will go to the other library instead of the Home page

* Added is present check

* "Fail Fast" fix

* Revert ""Fail Fast" fix"

This reverts commit 144b7821df03a3d84532228409e2a3faab2f7b8d.