feat: add method to convert Cygwin-style paths to Windows format
Created the method convertCygwinPathToWindows in the FileUtil class. It transforms a file path from Cygwin-style to Windows-style only if the operating system is Windows.

The method follows this logic:
1. If the filePath is null or the OS is not Windows, return it unchanged.
2. If the path starts with /cygdrive/{drive} or /{drive}, replace it with {driveLetterUpperCase}:\ and convert forward slashes to backslashes.

Based on research, the two common Cygwin-style formats are:
- /cygdrive/{drive}/some/path
- /c/some/path

If none of theses patters match, or if the OS is not Windows, the original path is returned.

Additionally, unit tests for this method were added to FileUtilTest class

OBS: I need help to run the second test when Windows is disabled... I am getting an error because the test is not recognized...