Fixes #13274: Allow cygwin-paths on Windows (#13297)
* feat: add method to convert Cygwin-style paths to Windows format

Created the method convertCygwinPathToWindows in the FileUtil class. It transforms a file path from Cygwin-style to Windows-style only if the operating system is Windows.

The method follows this logic:
1. If the filePath is null or the OS is not Windows, return it unchanged.
2. If the path starts with /cygdrive/{drive} or /{drive}, replace it with {driveLetterUpperCase}:\ and convert forward slashes to backslashes.

Based on research, the two common Cygwin-style formats are:
- /cygdrive/{drive}/some/path
- /c/some/path

If none of theses patters match, or if the OS is not Windows, the original path is returned.

Additionally, unit tests for this method were added to FileUtilTest class

OBS: I need help to run the second test when Windows is disabled... I am getting an error because the test is not recognized...

* feat: apply the convertCygwinPathToWinds on the path to prevent the ImportException

I need help to understand if I really need to do this change... it was not clear to me on the issue description. (it is my first open source contribuition)

* fix: remove FileUtil method

Do not know if I need to do that.

* feat: add Cygwin file path /mtn/

* feat: add test to Cygwin file path /mtn/

* feat: add entry on CHANGELOG.md and weave the method into JabKit.

* feat: remove convertCygwinPathToWindows

* feat: create converters and add new package to java modules

* feat: apply converters to --inputs and --outputs

* style: remove blank space

* feat: change the DataType from String and File to Path

* feat: remove code duplication

* fix: correct the style

* feat: create a new method to remove code duplication, create constants that represents the cygwin-paths known and the /c/ pattern

* feat: correct the unit tests to the changes made in the class

* Fix JBang

* Remove unused import

---------

Co-authored-by: Oliver Kopp <kopp.dev@gmail.com>