Fixes Bug for required field check
When a BibEntry type has <= 1 entries or the entries had no differing
fields, the required field check was skipped. This is a bug solved by
removing the if statement with a continue body in
BibliographyConsistencyCheck.check. This commit also refactors 2 tests.
The tests would otherwise fail since some entries they defined lacked
required fields and previously relied on a lack of required fields check
to pass